+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 # 0.3.1 - 2023-11-06
 
 ### Fixed
-- Fix bug #17
+- Fix random error at startup
 
 # 0.3.0 - 2023-10-30