+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Added
 - Average divergence
 - Allow cancelling screenshots
+- `move_dist` command
+- Value control commands without args now print out the current value
 
 # 0.3.1 - 2023-11-06